Change group label's numeric type from int64_t to float.
Previously, by group label, it could only filter numeric value by int
type, such as 1, 2, etc, now it's revised to enable filter by float type
such as 1.5, 2.5, etc.

+++ source/core/mining-manager/group-manager/GroupCounterLabelBuilder.cpp
@@ -81,21 +81,21 @@ bool convertRangeLabel(const std::string& src, NumericRangeGroupLabel::NumericRa
         return false;
     }
 
-    int64_t lowerBound = std::numeric_limits<int64_t>::min();
-    int64_t upperBound = std::numeric_limits<int64_t>::max();
+    float lowerBound = std::numeric_limits<float>::min();
+    float upperBound = std::numeric_limits<float>::max();
 
     try
     {
         if (delimitPos)
         {
             std::string sub = src.substr(0, delimitPos);
-            lowerBound = boost::lexical_cast<int64_t>(sub);
+            lowerBound = boost::lexical_cast<float>(sub);
         }
 
         if (delimitPos+1 != src.size())
         {
             std::string sub = src.substr(delimitPos+1);
-            upperBound = boost::lexical_cast<int64_t>(sub);
+            upperBound = boost::lexical_cast<float>(sub);
         }
     }
     catch(const boost::bad_lexical_cast& e)
